Move Serializer to folder
