+++ SchemeCreator.csproj
@@ -150,7 +150,9 @@
     <Compile Include="Data\Models\SmartGrid.cs" />
     <Compile Include="Data\Services\Colorer.cs" />
     <Compile Include="Data\Services\GateWorkPatterns.cs" />
-    <Compile Include="Data\Services\Serializer.cs" />
+    <Compile Include="Data\Services\Serialization\Gate.cs" />
+    <Compile Include="Data\Services\Serialization\Serializer.cs" />
+    <Compile Include="Data\Services\Serialization\Wire.cs" />
     <Compile Include="Data\Services\TextController.cs" />
     <Compile Include="Data\Services\Tracer.cs" />
     <Compile Include="Data\Models\Wire.cs" />
